requirements.txt
requests==2.32.2
python-dotenv==1.0.0
kivy==2.3.0
kivymd==1.1.1
pyinstaller==6.5.0 